Update Join Staff list to be responsive automatically
