+++ app/javascript/components/src/views/JoinStaffList/components/JoinStaffCard/index.jsx
@@ -4,80 +4,80 @@ import PropTypes from 'prop-types';
 import './styles/index.scss';
 
 const CancelButton = () => {
-	return(
-		<a href="#" className="cancel-button" >
+  return (
+    <a href="#" className="cancel-button" >
       <span className="icon has-text-grey">
-				<i className="fas fa-minus-circle"></i>
-			</span>
+        <i className="fas fa-minus-circle"></i>
+      </span>
     </a>
-	);
+  );
 }
 
-export const JoinStaffCard = ({id, name, phoneNumber, email, tickets, tenants, admin}) => {
-	
-	const InfoLabel = ({type}) => {
-		if (type === "tickets"){
-			return (
-				<div className="info-font">
+export const JoinStaffCard = ({ id, name, phoneNumber, email, tickets, tenants, admin }) => {
+
+  const InfoLabel = ({ type }) => {
+    if (type === "tickets") {
+      return (
+        <div className="info-font">
           <p className="number-font" >{tickets}</p> Open Tickets
         </div>
-			);
-		} else if (type === "tenants") {
-			return (
-				<div className="info-font">
+      );
+    } else if (type === "tenants") {
+      return (
+        <div className="info-font">
           <p className="number-font" >{tenants}</p> Tenants
         </div>
-			);
-		}
-	}
+      );
+    }
+  }
 
-	const AdminLabel = () => {
-		return (
-			<div className="admin-label"> Admin </div>
-		);
-	}
+  const AdminLabel = () => {
+    return (
+      <div className="admin-label"> Admin </div>
+    );
+  }
 
-	const tinyLinkSpacing = !(tickets && tenants) ? "tiny-link" : "tiny-link-info";
+  const tinyLinkSpacing = !(tickets && tenants) ? "tiny-link" : "tiny-link-info";
 
-	return (
-		<>
-			<div className="card card-length">
-				<div className="name-spacing">
-					<Link to={`/manage/staff/${id}`}>
-						<p className="name-font">
-							{name}
-						</p>
-					</Link>
-					{!admin && <CancelButton />}
-				</div>
-				<div className="card-content">
-					{admin && <AdminLabel />}
-				    <p className="info-font">
-				    	{phoneNumber}
-				    </p>
-				    <p className="info-font">
-				    	{email}
-				    </p>
-				    <hr className="card-divider" ></hr>
-				    {<InfoLabel type="tickets" />}
-				    {<InfoLabel type="tenants" />}
-						<Link to={`/manage/staff/${id}`}>
-							<p className={tinyLinkSpacing}>
-								Reassign
-							</p>
-						</Link>
-				</div>
-			</div>
-			<div className="card-separating-space"></div>
-		</>
-	);
+  return (
+    <>
+      <div className="card card-length">
+        <div className="name-spacing">
+          <Link to={`/manage/staff/${id}`}>
+            <p className="name-font">
+              {name}
+            </p>
+          </Link>
+          {!admin && <CancelButton />}
+        </div>
+        <div className="card-content">
+          {admin && <AdminLabel />}
+          <p className="info-font">
+            {phoneNumber}
+          </p>
+          <p className="info-font">
+            {email}
+          </p>
+          <hr className="card-divider" ></hr>
+          {<InfoLabel type="tickets" />}
+          {<InfoLabel type="tenants" />}
+          <Link to={`/manage/staff/${id}`}>
+            <p className={tinyLinkSpacing}>
+              Reassign
+            </p>
+          </Link>
+        </div>
+      </div>
+      <div className="card-separating-space"></div>
+    </>
+  );
 }
 
 JoinStaffCard.propTypes = {
-	name: PropTypes.string.isRequired,
-	phoneNumber: PropTypes.string.isRequired,
-	email: PropTypes.string.isRequired,
-	tickets: PropTypes.number,
-	tenants: PropTypes.number,
-	admin: PropTypes.bool,
+  name: PropTypes.string.isRequired,
+  phoneNumber: PropTypes.string.isRequired,
+  email: PropTypes.string.isRequired,
+  tickets: PropTypes.number,
+  tenants: PropTypes.number,
+  admin: PropTypes.bool,
 }
