+++ app/javascript/components/src/views/JoinStaffList/styles/index.scss
@@ -1,4 +1,5 @@
 @import "../../../scss/variables.scss";
+
 // Card
 .move-card {
 	padding-left: 400px;
